Fix Win64 build after r24896
TBR=mstarzinger@chromium.org

Review URL: https://codereview.chromium.org/646943007

Cr-Commit-Position: refs/heads/master@{#24898}
git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@24898 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/compiler/generic-node-inl.h
@@ -236,8 +236,7 @@ template <class B, class S>
 S* GenericNode<B, S>::New(GenericGraphBase* graph, int input_count, S** inputs,
                           bool has_extensible_inputs) {
   size_t node_size = sizeof(GenericNode);
-  size_t reserve_input_count =
-      has_extensible_inputs ? kDefaultReservedInputs : 0;
+  int reserve_input_count = has_extensible_inputs ? kDefaultReservedInputs : 0;
   size_t inputs_size = (input_count + reserve_input_count) * sizeof(Input);
   size_t uses_size = input_count * sizeof(Use);
   int size = static_cast<int>(node_size + inputs_size + uses_size);
